atualização: pagina lista de postagens e detalhes da postagem
